Added more casts, added 40/48 bit methods to main

+++ sacabench/saca/sais.hpp
@@ -223,7 +223,7 @@ public:
         container<ssize> SA = make_container<ssize>(text.size());
         run_saca<string_span>(text, SA, alphabet.max_character_value());
         for (size_t i = 0; i < text.size(); i++) {
-            DCHECK_LE(SA[i], std::numeric_limits<sa_index>::max());
+            DCHECK_LE(static_cast<size_t>(SA[i]), std::numeric_limits<sa_index>::max());
             out_sa[i] = static_cast<sa_index>(SA[i]);
         }
     }
